pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "podalize"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "3.6.1"
pyannote-audio = "3.2.0"
pyannote-core = "5.0.0"
altair = "<5"
pydub = "0.25.1"
pylatex = "1.4.1"
python-magic = "0.4.27"
pytube = "12.1.0"
streamlit = { extras = ["testing"], version = "^1.35.0" }
torchaudio = "^2.3.1"
openai-whisper = "20231117"
yt-dlp = "^2024.5.27"
wordcloud = "^1.9.3"
beartype = "^0.18.5"
pydantic = "^2.7.4"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.26.1"
ruff = "^0.4.8"
pytest = "^8.2.2"
pytest-sugar = "^1.0.0"
pytest-cov = "^5.0.0"
mypy = "^1.10.0"

[tool.poe.tasks]
_ruff_lint = "ruff check ."
_mypy_lint = "mypy ."
lint = ["_mypy_lint", "_ruff_lint"]
format = "ruff check --fix ."
test = "pytest"
app = "streamlit run podalize/__main__.py"

[tool.poe.tasks.debug]
cmd = "streamlit run podalize/__main__.py"
env = { LOG_LEVEL = "DEBUG" }

[tool.pytest.ini_options]
addopts = [
    "--cov=podalize",
    "--cov-report=term-missing",
    "--cov-report=xml:coverage.xml",
]

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # use of assert
    "S101",
]

fixable = ["ALL"]
unfixable = []
